Prepare for runtime log configuration
Several subdirs are needed under /var/log. To avoid corrupting the underlying
SD card /var/log is implemented as a tmpfs which is emptied on each boot.
To ensure the subdirectories of /var/log that various packages need will exist,
collate them into a tar file that gets unpacked after the tmpfs is mounted.