Use consistent filesystem label for volumio_data partition
scripts/x86image.sh uses a different filesystem label than all the other
platform setup scripts for the 'data' partition on the flash image.
To simplify the codebase, make x86 consistent with the other platforms.
he only place the label was used was in the udisks-glue.conf file, to which
it was only recently added (2db66874c2, Do not mount volumiodata partition).
This patch reverts that change as well.